Update script files and devide Push and others
