+++ ncmb_unity/Assets/NCMB/NCMBPush.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a4efb339912de4baa91fd57a0fd0f474
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
