+++ ncmb_unity/Assets/NCMB/NCMBPush/NCMBAnalyticsUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ca1230c8b68c646d48473b117c2f43b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
