+++ ncmb_unity/Assets/NCMB/NCMBPush/NCMBInstallationUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7781753b1226040f1af86832044a4d8c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
