+++ ncmb_unity/Assets/NCMB/NCMBPush/NCMBPushUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5d0975e790e7b4e0e97c22a184789503
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
