+++ ncmb_unity/Assets/NCMB/NCMBPushSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 98bf974f9df474d94abbe1caf875dd58
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
