+++ ncmb_unity/Assets/NCMB/Script/NCMBInstallation.cs
@@ -70,7 +70,7 @@ namespace  NCMB
 		/// コンストラクター。<br/>
 		/// JSONデータをセットしinstallationを作成する場合、こちらを使用します。
 		/// </summary>
-		internal NCMBInstallation (string jsonText): base ()//NCMBObjectのコンストラクタ実行
+		internal NCMBInstallation (string jsonText): base () //NCMBObjectのコンストラクタ実行
 		{
 			if (jsonText != null && jsonText != "") {
 				Dictionary<string, object> dic = Json.Deserialize (jsonText) as Dictionary<string, object>;	//辞書形式へ変換
@@ -88,7 +88,7 @@ namespace  NCMB
 			DeviceToken = NCMBManager._token;
 			//applicationName,appVersion,deviceType,timeZone,SdkVersionを取得/設定
 			#if !UNITY_EDITOR
-			setDefaultProperty ();
+			SetDefaultProperty ();
 			#endif
 		}
 
@@ -125,33 +125,33 @@ namespace  NCMB
 			}
 		}
 
-		/// <summary>
-		/// デバイストークンの取得を行います。 <br/>
-		/// 通信結果が必要な場合はコールバックを指定するこちらを使用します。
-		/// </summary>
-		/// <param name="callback">コールバック</param>
-		public void GetDeviceToken(NCMBGetCallback<String> callback){
-			if(this.ContainsKey("deviceToken") && this["deviceToken"] != null ){
-				callback((string)this["deviceToken"], null);
-			} else {
-				new Thread(() => {
-					for (int i = 0; i < 10; i++){
-						if (NCMBManager._token != null){
-							this["deviceToken"] = NCMBManager._token;
-							break;
-						}
-						Thread.Sleep(500);
-					}
-					if (callback != null){
-						if (this.ContainsKey("deviceToken") && this["deviceToken"] != null){
-							callback((string)this["deviceToken"], null);
-						} else {
-							callback(null, new NCMBException("Can not get device token"));
-						}
-					}
-				}).Start();
-			}
-		}
+		// /// <summary>
+		// /// デバイストークンの取得を行います。 <br/>
+		// /// 通信結果が必要な場合はコールバックを指定するこちらを使用します。
+		// /// </summary>
+		// /// <param name="callback">コールバック</param>
+		// public void GetDeviceToken(NCMBGetCallback<String> callback){
+		// 	if(this.ContainsKey("deviceToken") && this["deviceToken"] != null ){
+		// 		callback((string)this["deviceToken"], null);
+		// 	} else {
+		// 		new Thread(() => {
+		// 			for (int i = 0; i < 10; i++){
+		// 				if (NCMBManager._token != null){
+		// 					this["deviceToken"] = NCMBManager._token;
+		// 					break;
+		// 				}
+		// 				Thread.Sleep(500);
+		// 			}
+		// 			if (callback != null){
+		// 				if (this.ContainsKey("deviceToken") && this["deviceToken"] != null){
+		// 					callback((string)this["deviceToken"], null);
+		// 				} else {
+		// 					callback(null, new NCMBException("Can not get device token"));
+		// 				}
+		// 			}
+		// 		}).Start();
+		// 	}
+		// }
 
 		/// <summary>
 		/// Android/iOSの取得、または設定を行います。
