+++ ncmb_unity/Assets/NCMB/Script/NCMBPush.cs
@@ -29,72 +29,80 @@ namespace NCMB
 	[NCMBClassName ("push")]
 	public class NCMBPush:NCMBObject
 	{
-		#if UNITY_ANDROID
-		static AndroidJavaClass m_AJClass;
-
-#elif UNITY_IOS
-		[DllImport ("__Internal")]
-		private static extern void registerNotification (bool useAnalytics);
-
-		[DllImport ("__Internal")]
-		private static extern void registerNotificationWithLocation ();
-
-		[DllImport ("__Internal")]
-		private static extern void clearAll ();
-		#endif
-		/*		** 初期化 ***/
-		static NCMBPush ()
-		{
-			#if UNITY_ANDROID && !UNITY_EDITOR
-                        m_AJClass = new AndroidJavaClass("com.nifcloud.mbaas.ncmbfcmplugin.FCMInit");
-			#endif
-
-		}
-
+// 		#if UNITY_ANDROID
+// 		static AndroidJavaClass m_AJClass;
+//
+// #elif UNITY_IOS
+// 		[DllImport ("__Internal")]
+// 		private static extern void registerNotification (bool useAnalytics);
+//
+// 		[DllImport ("__Internal")]
+// 		private static extern void registerNotificationWithLocation ();
+//
+// 		[DllImport ("__Internal")]
+// 		private static extern void clearAll ();
+// 		#endif
+// 		/*		** 初期化 ***/
+// 		static NCMBPush ()
+// 		{
+// 			#if UNITY_ANDROID && !UNITY_EDITOR
+//                         m_AJClass = new AndroidJavaClass("com.nifcloud.mbaas.ncmbfcmplugin.FCMInit");
+// 			#endif
+//
+// 		}
+//
 		/// <summary>
 		/// コンストラクター。<br/>
 		/// プッシュの作成を行います。
 		/// </summary>
 		public NCMBPush () : base ()	//継承元のコンストラクタを実施するため
 		{
-		}
-		#if UNITY_ANDROID
-		public static void Register ()
-		{
+			if (Type.GetType("NCMBPushUtils") != null) {
+				//If Push plugin is imported
+				NCMBPushUtils pushUtil = new NCMBPushUtils();
+			} else {
+				//If Push plugin is not imported
 
-#if !UNITY_EDITOR
-				m_AJClass.CallStatic("Init");
-		#endif
+			}
 		}
+// 		#if UNITY_ANDROID
+// 		public static void Register ()
+// 		{
+//
+// #if !UNITY_EDITOR
+// 				m_AJClass.CallStatic("Init");
+// 		#endif
+// 		}
+//
+// #elif UNITY_IOS
+// 		public static void Register (bool useAnalytics)
+// 		{
+// 			#if !UNITY_EDITOR
+// 		registerNotification(useAnalytics);
+// 			#endif
+// 		}
+// 		#endif
+// 		#if UNITY_ANDROID
+// 	public static void RegisterWithLocation ()
+// 	{
+//
+// #if !UNITY_EDITOR
+// 			m_AJClass.CallStatic("Init");
+// 		#endif
+// 	}
+//
+// #elif UNITY_IOS
+// 		/// <summary>
+// 		/// Register for receiving remote notifications (with current location).
+// 		/// </summary>
+// 		internal static void RegisterWithLocation ()
+// 		{
+// 			#if !UNITY_EDITOR
+// 	registerNotificationWithLocation();
+// 			#endif
+// 		}
+// 		#endif
 
-#elif UNITY_IOS
-		public static void Register (bool useAnalytics)
-		{
-			#if !UNITY_EDITOR
-		registerNotification(useAnalytics);
-			#endif
-		}
-		#endif
-		#if UNITY_ANDROID
-	public static void RegisterWithLocation ()
-	{
-
-#if !UNITY_EDITOR
-			m_AJClass.CallStatic("Init");
-		#endif
-	}
-
-#elif UNITY_IOS
-		/// <summary>
-		/// Register for receiving remote notifications (with current location).
-		/// </summary>
-		internal static void RegisterWithLocation ()
-		{
-			#if !UNITY_EDITOR
-	registerNotificationWithLocation();
-			#endif
-		}
-		#endif
 		/*** Push設定 ***/
 		/// <summary>
 		/// メッセージの取得、または設定を行います。
@@ -327,24 +335,24 @@ namespace NCMB
 			base.SaveAsync (callback);
 		}
 
-		#region Process notification for iOS only
-
-		#if UNITY_IOS
-		// Clears all notifications.
-		public void ClearAll ()
-		{
-			#if UNITY_4_0 || UNITY_4_1 || UNITY_4_2 || UNITY_4_3 || UNITY_4_5 || UNITY_4_6 || UNITY_4_7
-			NotificationServices.ClearRemoteNotifications ();
-			#else
-			UnityEngine.iOS.NotificationServices.ClearRemoteNotifications ();
-			#endif
-
-			#if !UNITY_EDITOR
-		clearAll();
-			#endif
-		}
-		#endif
-		#endregion
+		// #region Process notification for iOS only
+		//
+		// #if UNITY_IOS
+		// // Clears all notifications.
+		// public void ClearAll ()
+		// {
+		// 	#if UNITY_4_0 || UNITY_4_1 || UNITY_4_2 || UNITY_4_3 || UNITY_4_5 || UNITY_4_6 || UNITY_4_7
+		// 	NotificationServices.ClearRemoteNotifications ();
+		// 	#else
+		// 	UnityEngine.iOS.NotificationServices.ClearRemoteNotifications ();
+		// 	#endif
+		//
+		// 	#if !UNITY_EDITOR
+		// clearAll();
+		// 	#endif
+		// }
+		// #endif
+		// #endregion
 
 		/// <summary>
 		/// Push内のオブジェクトで使用出来るクエリを取得します。
